Fx x64 Warnings

+++ TRXMapAreas.cpp
@@ -422,7 +422,7 @@ BOOL CTRXMapAreas::OnInitDialog()
 		int y = LevelAreas [ i ].rect.top;
 		int cx = LevelAreas [ i ].rect.right - LevelAreas [ i ].rect.left;
 		int cy = LevelAreas [ i ].rect.bottom - LevelAreas [ i ].rect.top;
-		LevelAreas [ i ].pWnd->SetWindowPos ( previousWIndow, x, y, cx, cy, (UINT)( HWND_TOP ) );
+		LevelAreas [ i ].pWnd->SetWindowPos ( previousWIndow, x, y, cx, cy, (UINT)(unsigned long long)( HWND_TOP ) );
 		previousWIndow	= LevelAreas [ i ].pWnd;
 
 		//
@@ -898,7 +898,7 @@ void CTRXMapAreas::ShowMap()
 		int y = LevelAreas [ i ].rect.top;
 		int cx = LevelAreas [ i ].rect.right - LevelAreas [ i ].rect.left;
 		int cy = LevelAreas [ i ].rect.bottom - LevelAreas [ i ].rect.top;
-		LevelAreas [ i ].pWnd->SetWindowPos ( previousWIndow, x, y, cx, cy, (UINT)( HWND_TOP ) );
+		LevelAreas [ i ].pWnd->SetWindowPos ( previousWIndow, x, y, cx, cy, (UINT)(unsigned long long)( HWND_TOP ) );
 		previousWIndow	= LevelAreas [ i ].pWnd;
 
 	}
